Header file for functions & bug fixes
functions.h will now have all the definitions
Fixed node not being removed when given index value was negative and evaluated to first node
